+++ Changelog.md
@@ -2,6 +2,7 @@
 
 **latest**
 - complete rewrite
+- `PSQL_TRUST_LOCALNET` config parameter renamed to `PG_TRUST_LOCALNET`
 
 **9.4-2**
 - added replication options